backend/requirements.txt
#!/usr/bin/env python
# Crypto Portfolio API Backend Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database & ORM
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.8
redis==5.0.1

# Environment & Configuration
python-dotenv==1.0.0
pydantic-settings==2.1.0

# HTTP Client & Requests
httpx==0.25.2
requests==2.31.0
aiofiles==23.2.1

# Date & Time Utilities
python-dateutil==2.8.2
pytz==2023.3

# External APIs
python-binance==1.0.19

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Email validation (for pydantic EmailStr)
email-validator==2.1.0